countries and Argentina is out
-removed Argentina from check
-organized into src folder
-some variable name changes
-added country
